can't leave an event if balance!=0

+++ client/src/main/java/client/scenes/SplittyOverviewCtrl.java
@@ -110,8 +110,6 @@ public class SplittyOverviewCtrl implements Initializable {
                 }
             }
         });
-//        fetchExpenses();
-//        fetchParticipants();
     }
 
     public void setEventCode(int eventCode){
@@ -227,8 +225,9 @@ public class SplittyOverviewCtrl implements Initializable {
         expenseList.getItems().addAll(expenses);
         allExpenses.setContent(expenseList);
     }
-    public void fetchParticipants(){
 
+    public void fetchParticipants(){
+        participantListView.getItems().clear();
         List<Participant> participants = new ArrayList<>();
         try{
             participants = serverUtils.getParticipants(eventCode);
@@ -283,6 +282,17 @@ public class SplittyOverviewCtrl implements Initializable {
     }
 
     public void leaveEvent(ActionEvent actionEvent) {
+        // can only leave if balance is 0
+        Participant me = serverUtils.getParticipant(mainCtrl.getMyUuid(), eventCode);
+        if (me.getBalance() != 0) {
+            Alert alert = new Alert(Alert.AlertType.ERROR);
+            alert.setTitle("Error");
+            alert.setHeaderText("You can't leave the event");
+            alert.setContentText("You can only leave the event if your balance is 0");
+            alert.showAndWait();
+            return;
+        }
+
         serverUtils.deleteParticipant(mainCtrl.getMyUuid(), eventCode);
         mainCtrl.showStartScreen();
         confirmationLabel.setVisible(false);
@@ -336,6 +346,10 @@ public class SplittyOverviewCtrl implements Initializable {
         confirmationLabel.setVisible(false);
         cancelLeaveButton.setVisible(false);
         leaveConfirmationButton.setVisible(false);
+    }
+
+    public int getCurrentEventCode(){
+        return this.eventCode;
     }
 }
